+++ IdentityServer/v7/Configuration/Permissions/ConsoleDcrClient/ConsoleDcrClient.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="IdentityModel" Version="7.0.0-preview.1" />
+      <PackageReference Include="IdentityModel" Version="7.0.0" />
     </ItemGroup>
 
 </Project>
