+++ IdentityServer/v7/Configuration/SimpleDcr/Configuration/Configuration.csproj
@@ -7,10 +7,10 @@
   </PropertyGroup>
   
   <ItemGroup>
-    <PackageReference Include="Duende.IdentityServer.Configuration" Version="7.0.0" />
-    <PackageReference Include="Duende.IdentityServer.Configuration.EntityFramework" Version="7.0.0" />
-    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.1" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.1" />
+    <PackageReference Include="Duende.IdentityServer.Configuration" Version="7.0.5" />
+    <PackageReference Include="Duende.IdentityServer.Configuration.EntityFramework" Version="7.0.5" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="8.0.6" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Sqlite" Version="8.0.6" />
   </ItemGroup>
 
 </Project>
