+++ IdentityServer/v7/Configuration/SimpleDcr/ConsoleDcrClient/ConsoleDcrClient.csproj
@@ -7,7 +7,7 @@
     </PropertyGroup>
 
     <ItemGroup>
-      <PackageReference Include="IdentityModel" Version="6.2.0" />
+      <PackageReference Include="IdentityModel" Version="7.0.0" />
     </ItemGroup>
 
 </Project>
