--- IdentityServer/v7/Configuration/SimpleDcr/SimpleApi/Startup.cs
@@ -1,42 +1,0 @@
-using System.IdentityModel.Tokens.Jwt;
-using Microsoft.AspNetCore.Builder;
-using Microsoft.Extensions.DependencyInjection;
-
-namespace SimpleApi
-{
-    public class Startup
-    {
-        public void ConfigureServices(IServiceCollection services)
-        {
-            services.AddControllers();
-
-            // this API will accept any access token from the authority
-            services.AddAuthentication("token")
-                .AddJwtBearer("token", options =>
-                {
-                    options.Authority = "https://localhost:5001";
-                    options.MapInboundClaims = false;
-                    
-                    options.TokenValidationParameters.ValidateAudience = false;
-                    options.TokenValidationParameters.ValidTypes = new[] { "at+jwt" };
-                });
-
-            services.AddAuthorization(options =>
-            {
-                options.AddPolicy("SimpleApi", p => p.RequireClaim("scope", "SimpleApi"));
-            });
-        }
-
-        public void Configure(IApplicationBuilder app)
-        {
-            app.UseRouting();
-            app.UseAuthentication();
-            app.UseAuthorization();
-
-            app.UseEndpoints(endpoints =>
-            {
-                endpoints.MapControllers().RequireAuthorization();
-            });
-        }
-    }
-}
