+++ IdentityServer/v7/DPoP/Api/DPoP/ConfigureJwtBearerOptions.cs
@@ -1,6 +1,5 @@
 using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.Extensions.Options;
-using System;
 
 namespace Api;
