+++ IdentityServer/v7/DPoP/Api/DPoP/DPoPExtensions.cs
@@ -1,10 +1,7 @@
 using IdentityModel;
 using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Http;
 using Microsoft.IdentityModel.Tokens;
-using System.Collections.Generic;
 using System.Diagnostics.CodeAnalysis;
-using System.Linq;
 using System.Text.Json;
 
 namespace Api;
