+++ IdentityServer/v7/DPoP/Api/DPoP/DPoPOptions.cs
@@ -1,6 +1,4 @@
-﻿using System;
-
-namespace Api;
+﻿namespace Api;
 
 public class DPoPOptions
 {
