+++ IdentityServer/v7/DPoP/Api/DPoP/DPoPProofValidatonContext.cs
@@ -1,6 +1,3 @@
-using System.Collections.Generic;
-using System.Security.Claims;
-
 namespace Api;
 
 public class DPoPProofValidatonContext
