+++ IdentityServer/v7/DPoP/Api/DPoP/DPoPServiceCollectionExtensions.cs
@@ -1,7 +1,5 @@
 using Microsoft.AspNetCore.Authentication.JwtBearer;
-using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Options;
-using System;
 
 namespace Api;
