+++ IdentityServer/v7/DPoP/Api/DPoP/IReplayCache.cs
@@ -1,6 +1,3 @@
-using System;
-using System.Threading.Tasks;
-
 namespace Api;
 
 public interface IReplayCache
