--- IdentityServer/v7/DPoP/Api/Startup.cs
@@ -1,36 +1,0 @@
-namespace Api;
-
-public class Startup
-{
-    public void ConfigureServices(IServiceCollection services)
-    {
-        services.AddControllers();
-        services.AddCors();
-
-        // this API will accept any access token from the authority
-        services.AddAuthentication("token")
-            .AddJwtBearer("token", options =>
-            {
-                options.Authority = "https://localhost:5001";
-                options.TokenValidationParameters.ValidateAudience = false;
-                options.MapInboundClaims = false;
-
-                options.TokenValidationParameters.ValidTypes = new[] { "at+jwt" };
-            });
-
-        // layers DPoP onto the "token" scheme above
-        services.ConfigureDPoPTokensForScheme("token");
-    }
-
-    public void Configure(IApplicationBuilder app)
-    {
-        app.UseRouting();
-        app.UseAuthentication();
-        app.UseAuthorization();
-
-        app.UseEndpoints(endpoints =>
-        {
-            endpoints.MapControllers().RequireAuthorization();
-        });
-    }
-}
