+++ IdentityServer/v7/DPoP/WebClient/Controllers/HomeController.cs
@@ -1,7 +1,5 @@
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Net.Http;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Authentication;
 using Duende.AccessTokenManagement.OpenIdConnect;
