+++ IdentityServer/v7/DPoP/WebClient/Program.cs
@@ -1,45 +1,103 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using Microsoft.AspNetCore.Authentication;
+using Microsoft.IdentityModel.Tokens;
 using Serilog;
-using Serilog.Events;
-using System;
+using System.Security.Cryptography;
+using System.Text.Json;
 
-namespace WebClient;
+Console.Title = "WebClient";
 
-public class Program
+Log.Logger = new LoggerConfiguration()
+    .MinimumLevel.Information()
+    .Enrich.FromLogContext()
+    .WriteTo.Console(outputTemplate: "[{Timestamp:HH:mm:ss} {Level}] {SourceContext}{NewLine}{Message:lj}{NewLine}{Exception}{NewLine}")
+    .CreateLogger();
+
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddSerilog();
+builder.Services.AddControllersWithViews();
+
+// add cookie-based session management with OpenID Connect authentication
+builder.Services.AddAuthentication(options =>
 {
-    public static int Main(string[] args)
+    options.DefaultScheme = "cookie";
+    options.DefaultChallengeScheme = "oidc";
+})
+    .AddCookie("cookie", options =>
     {
-        Console.Title = "WebClient";
+        options.Cookie.Name = "dpop";
 
-        Log.Logger = new LoggerConfiguration()
-            .MinimumLevel.Information()
-            .Enrich.FromLogContext()
-            .WriteTo.Console(outputTemplate: "[{Timestamp:HH:mm:ss} {Level}] {SourceContext}{NewLine}{Message:lj}{NewLine}{Exception}{NewLine}")
-            .CreateLogger();
+        options.ExpireTimeSpan = TimeSpan.FromHours(8);
+        options.SlidingExpiration = false;
 
-        try
+        options.Events.OnSigningOut = async e =>
         {
-            Log.Information("Starting host...");
-            CreateHostBuilder(args).Build().Run();
-            return 0;
-        }
-        catch (Exception ex)
-        {
-            Log.Fatal(ex, "Host terminated unexpectedly.");
-            return 1;
-        }
-        finally
-        {
-            Log.CloseAndFlush();
-        }
-    }
+            // automatically revoke refresh token at signout time
+            await e.HttpContext.RevokeRefreshTokenAsync();
+        };
+    })
+    .AddOpenIdConnect("oidc", options =>
+    {
+        options.Authority = "https://localhost:5001";
+        options.RequireHttpsMetadata = false;
 
-    public static IHostBuilder CreateHostBuilder(string[] args) =>
-        Host.CreateDefaultBuilder(args)
-            .ConfigureWebHostDefaults(webBuilder =>
-            {
-                webBuilder.UseStartup<Startup>();
-            })
-            .UseSerilog();
-}
+        options.ClientId = "dpop";
+        options.ClientSecret = "905e4892-7610-44cb-a122-6209b38c882f";
+
+        // code flow + PKCE (PKCE is turned on by default)
+        options.ResponseType = "code";
+        options.ResponseMode = "query";
+        options.UsePkce = true;
+
+        options.Scope.Clear();
+        options.Scope.Add("openid");
+        options.Scope.Add("profile");
+        options.Scope.Add("scope1");
+        options.Scope.Add("offline_access");
+
+        // keeps id_token smaller
+        options.GetClaimsFromUserInfoEndpoint = true;
+        options.SaveTokens = true;
+        options.MapInboundClaims = false;
+        options.DisableTelemetry = true;
+
+        options.TokenValidationParameters = new TokenValidationParameters
+        {
+            NameClaimType = "name",
+            RoleClaimType = "role"
+        };
+    });
+
+// this is only needed if you want to override/prevent dpop from being used on certain API endpoints
+//services.AddTransient<IDPoPProofService, CustomProofService>();
+
+// add automatic token management
+builder.Services.AddOpenIdConnectAccessTokenManagement(options =>
+{
+    // create and configure a DPoP JWK
+    var rsaKey = new RsaSecurityKey(RSA.Create(2048));
+    var jwk = JsonWebKeyConverter.ConvertFromSecurityKey(rsaKey);
+    jwk.Alg = "PS256";
+    options.DPoPJsonWebKey = JsonSerializer.Serialize(jwk);
+});
+
+// add HTTP client to call protected API
+builder.Services.AddUserAccessTokenHttpClient("client", configureClient: client =>
+{
+    client.BaseAddress = new Uri("https://localhost:5005");
+});
+
+var app = builder.Build();
+
+app.UseDeveloperExceptionPage();
+app.UseHttpsRedirection();
+app.UseStaticFiles();
+
+app.UseRouting();
+
+app.UseAuthentication();
+app.UseAuthorization();
+
+app.MapDefaultControllerRoute().RequireAuthorization();
+
+app.Run();
