+++ IdentityServer/v7/DPoP/WebClient/TestHandler.cs
@@ -1,9 +1,3 @@
-using Microsoft.Extensions.Logging;
-using System.Linq;
-using System.Net.Http;
-using System.Threading;
-using System.Threading.Tasks;
-
 namespace WebClient;
 
 public class TestHandler : DelegatingHandler
